Move DepositService logic inside blockchain
